Feat stats

+++ src/files/files.controller.ts
@@ -8,21 +8,23 @@ import {
   UploadedFile,
   UseGuards,
   UseInterceptors,
+  NotFoundException,
 } from '@nestjs/common';
-import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiConsumes, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { FilesService } from './files.service';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { JwtAuthGuard } from '../auth/jwt.strategy';
 
 interface AuthRequest extends Request {
   user: {
-    id: string;
+    sub: string;
     username: string;
-    role: string;
+    isAdmin: boolean;
   };
 }
 
 @ApiTags('files')
+@ApiBearerAuth()
 @Controller('files')
 @UseGuards(JwtAuthGuard)
 export class FilesController {
@@ -30,11 +32,14 @@ export class FilesController {
 
   @Post('upload')
   @UseInterceptors(FileInterceptor('file'))
+  @ApiConsumes('multipart/form-data')
   @ApiOperation({ summary: 'Upload a file' })
-  @ApiResponse({ status: 201, description: 'File successfully uploaded' })
-  @ApiResponse({ status: 409, description: 'File already exists' })
+  @ApiResponse({ status: 201, description: 'File uploaded successfully', type: String })
   uploadFile(@UploadedFile() file: Express.Multer.File, @Req() req: AuthRequest): Promise<string> {
-    const userId = req.user.id;
+    if (!req.user || !req.user.sub) {
+      throw new NotFoundException('User authentication failed or user ID is missing');
+    }
+    const userId = req.user.sub;
     return this.filesService.uploadFile(file, userId);
   }
 
@@ -42,7 +47,10 @@ export class FilesController {
   @ApiOperation({ summary: 'Get a file' })
   @ApiResponse({ status: 200, description: 'File successfully retrieved' })
   getFile(@Param('key') fileKey: string, @Req() req: AuthRequest): Promise<string> {
-    const userId = req.user.id;
+    if (!req.user || !req.user.sub) {
+      throw new NotFoundException('User authentication failed or user ID is missing');
+    }
+    const userId = req.user.sub;
     return this.filesService.getFile(fileKey, userId);
   }
 
@@ -50,7 +58,10 @@ export class FilesController {
   @ApiOperation({ summary: 'Delete a file' })
   @ApiResponse({ status: 204, description: 'File deleted successfully' })
   deleteFile(@Param('key') key: string, @Req() req: AuthRequest): Promise<void> {
-    const userId = req.user.id;
+    if (!req.user || !req.user.sub) {
+      throw new NotFoundException('User authentication failed or user ID is missing');
+    }
+    const userId = req.user.sub;
     return this.filesService.delete(key, userId);
   }
 }
